APP/requirements.txt
flask==2.2.3
mysql-connector-python == 9.2.0
Werkzeug==2.2.3
